+++ components/common/bottom-sheet/bottom-sheet.type.ts
@@ -1,0 +1,32 @@
+import type { SelectOption } from '@/components/common/select/select.type';
+
+export interface BottomSheetProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  title: string;
+  description?: string;
+  children: React.ReactNode;
+}
+
+interface SelectBottomSheetSharedProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  title: string;
+  description?: string;
+  options: SelectOption[];
+  onConfirm: () => void;
+}
+
+interface SingleSelectBottomSheetProps extends SelectBottomSheetSharedProps {
+  multiple?: false;
+  value?: string;
+  onChange?: (value: string) => void;
+}
+
+interface MultiSelectBottomSheetProps extends SelectBottomSheetSharedProps {
+  multiple: true;
+  value?: string[];
+  onChange?: (value: string[]) => void;
+}
+
+export type SelectBottomSheetProps = SingleSelectBottomSheetProps | MultiSelectBottomSheetProps;
