+++ components/common/bottom-sheet/index.ts
@@ -1,0 +1,3 @@
+export { BottomSheet } from './BottomSheet';
+export { SelectBottomSheet } from './SelectBottomSheet';
+export type { BottomSheetProps, SelectBottomSheetProps } from './bottom-sheet.type';
